% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_unique.R
\name{count_unique}
\alias{count_unique}
\title{Count the number of unique values}
\usage{
count_unique(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
An integer giving the number of unique values in \code{x}.
}
\description{
Count the number of unique values
}
\keyword{internal}
