% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iopr-package.R
\docType{package}
\name{iopr-package}
\alias{iopr}
\alias{iopr-package}
\title{iopr: Inequality of Opportunity Analysis for Complex Surveys}
\description{
Provides functions to estimate Inequality of Opportunity (IOP) and the Gini index for complex survey designs. Supports stratified and clustered sampling with variance estimation using analytical (influence function) and bootstrap methods.
}
\author{
\strong{Maintainer}: Francis Bilson Darku \email{fbd17@yahoo.com}

Authors:
\itemize{
  \item Francis Bilson Darku \email{fbd17@yahoo.com}
}

}
\keyword{internal}
